+++ src/cloudant/query.py
@@ -182,12 +182,6 @@ class Query(dict):
                 'Add a selector to define the query and retry.'
             )
             raise CloudantArgumentError(msg)
-        if data.get('fields', None) is None or data.get('fields') == []:
-            msg = (
-                'No fields list in the query or the fields list was empty.  '
-                'Add a list of fields for the query and retry.'
-            )
-            raise CloudantArgumentError(msg)
 
         # Execute query find
         headers = {'Content-Type': 'application/json'}